# docs/api/openapi_spec.yaml
openapi: 3.0.3
info:
  title: HouseBroker Perú - API
  description: |
    Contrato de API REST de HouseBroker Perú (Django REST Framework).

    Este documento es **el contrato acumulado del proyecto**: cada sprint
    incorpora aquí los endpoints que entrega, de modo que al cierre de la
    carrera (Sprint 7) contenga la API completa sin fragmentarse en varios
    archivos.

    ### Contratos incorporados
    | Sprint | Tarea | Alcance |
    |---|---|---|
    | Sprint 1 | `TASK-ARC-PROP-01` | CRUD de propiedades (HU-PROP-01) |
    | Sprint 2 | `TASK-ARC-PROP-02` | Catálogo paginado `GET /api/v1/properties` (HU-PROP-02) |
    | Sprint 2 | `TASK-ARC-PROP-03` | Filtros de búsqueda en `GET /api/v1/properties` (HU-PROP-03) |
    | Sprint 3 | `TASK-ARC-PROP-04` | Estado operativo y horarios de visita (HU-PROP-04) |
    | Sprint 3 | `TASK-ARC-SEC-01` | Registro, inicio de sesión y refresco de token (HU-SEC-01) |
    | Sprint 3 | `TASK-ARC-SEC-02` | Recuperación de contraseña con token temporal (HU-SEC-02) |
    | Sprint 4 | `TASK-ARC-CRM-01` | Disponibilidad por fecha y creación de citas (HU-CRM-01) |
    | Sprint 4 | `TASK-ARC-CRM-02` | Agenda de citas filtrada por rol y estados de cita (HU-CRM-02) |
    | Sprint 4 | `TASK-ARC-PROP-05` | Gestión de propiedades favoritas por usuario (HU-PROP-05) |

    ### Convenciones
    - Todos los endpoints están bajo el prefijo `/api/v1`.
    - Los parámetros de paginación son opcionales; el servidor aplica
...
  version: 1.8.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/properties/{id}/management:
    parameters:
      - $ref: '#/components/parameters/PropertyId'

    get:
      tags:
        - Disponibilidad
      operationId: getManagedProperty
      summary: Obtener el detalle de gestión de una propiedad
      security:
        - bearerAuth: []
      description: |
        Devuelve el detalle del inmueble para el panel de disponibilidad.
        Solo el agente asignado o un administrador pueden acceder; un cliente
        o un agente ajeno reciben `403`, y la ausencia de sesión válida, `401`.

        No filtra por `is_active`: permite cargar inmuebles `SUSPENDIDO` para
        gestionar su disponibilidad y reactivarlos mediante `PATCH /api/v1/properties/{id}/status`.
        El detalle público sigue devolviendo `404` mientras el inmueble esté
        inactivo. En esta ruta, `404` indica que el inmueble no existe.
      responses:
        '200':
          description: Detalle de gestión de la propiedad, incluso si está suspendida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Propiedad no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemNotFound'
  /api/v1/properties/{id}/status:
    parameters:
      - $ref: '#/components/parameters/PropertyId'

    patch:
      tags:
        - Disponibilidad
      operationId: updatePropertyStatus
      summary: Actualizar el estado operativo de una propiedad
      description: |
        Cambia el estado operativo del inmueble (RF-PROP-04). Es un `PATCH`
        semántico: el cuerpo puede contener `status`, `reason` o ambos, y el
        resto de campos del inmueble queda intacto.

        `status` gobierna la visibilidad en el catálogo, no el borrado físico:

        | Estado | En catálogo | Agendable |
        |---|---|---|
        | `DISPONIBLE` | Sí | Sí |
        | `RESERVADO` | Sí | No |
        | `ALQUILADO` | Sí | No |
        | `VENDIDO` | Sí | No |
        | `SUSPENDIDO` | No (`is_active = false`) | No |

        La transición a `SUSPENDIDO` es la que retira el inmueble del catálogo
        mediante eliminación lógica. Volver a `DISPONIBLE` lo reactiva. Es
        intencionalmente distinta de `DELETE /api/v1/properties/{id}`, que
        también desactiva pero no deja registro de la suspensión.

        Desde `RESERVADO` se permite cerrar en `VENDIDO` con `reason` obligatorio;
        el inmueble sigue visible en catálogo, pero no acepta nuevas reservas.

        No hay transición válida desde `VENDIDO` ni `ALQUILADO`: son estados
        terminales y un intento de revertirlos produce `409`.

        El estado anterior se devuelve en `previous_status` para que el frontend
        pueda revertir la UI si observa un fallo de red después de aplicar el
        cambio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyStatusInput'
            examples:
              cambioDeEstado:
                summary: Marcar como reservada con motivo
                value:
                  status: RESERVADO
                  reason: El propietario inicio tramite de reserva con seña.
              soloEstado:
                summary: Reactivar un inmueble suspendido
                value:
                  status: DISPONIBLE
      responses:
        '200':
          description: Estado actualizado exitosamente
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyStatusResult'
              example:
                id: b2c1f5a0-1a2b-4c3d-9e0f-111111111111
                status: RESERVADO
                previous_status: DISPONIBLE
                is_active: true
                reason: El propietario inicio tramite de reserva con seña.
                changed_by: 7c4a9d21-3b6e-4f80-9a2d-5e7c1b3f8d40
                changed_at: '2026-09-30T14:20:00Z'
        '400':
          description: Cuerpo inválido o transición no permitida
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemBadRequest'
              examples:
                estadoInvalido:
                  summary: '`status` fuera del enum'
                  value:
                    type: https://housebroker.pe/errors/invalid-query-parameter
                    title: Parámetro de consulta inválido
                    status: 400
                    detail: El campo 'status' admite los valores DISPONIBLE, RESERVADO, ALQUILADO, VENDIDO, SUSPENDIDO.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/status
                    code: invalid_query_parameter
                    errors:
                      - parameter: status
                        message: El valor 'ALQUILANDO' no pertenece al conjunto permitido (DISPONIBLE, RESERVADO, ALQUILADO, VENDIDO, SUSPENDIDO).
                cuerpoVacio:
                  summary: 'PATCH sin `status`'
                  value:
                    type: https://housebroker.pe/errors/invalid-query-parameter
                    title: Parámetro de consulta inválido
                    status: 400
                    detail: El campo 'status' es obligatorio para actualizar el estado operativo.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/status
                    code: invalid_query_parameter
                    errors:
                      - parameter: status
                        message: Este campo es obligatorio.
                transicionInvalida:
                  summary: Transición desde un estado terminal
                  value:
                    type: https://housebroker.pe/errors/invalid-property-transition
                    title: Transición de estado no permitida
                    status: 400
                    detail: No se puede pasar de 'VENDIDO' a 'DISPONIBLE'. Los estados VENDIDO y ALQUILADO son terminales.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/status
                    code: invalid_property_transition
                    errors:
                      - parameter: status
                        message: La transición VENDIDO -> DISPONIBLE no está permitida.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Propiedad no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemNotFound'
        '409':
          description: |
            El inmueble tiene visitas confirmadas en la franja afectada, o la
            escritura concurrente que colisiona con otra actualización.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemConflict'
              example:
                type: https://housebroker.pe/errors/property-state-conflict
                title: Conflicto de estado del inmueble
                status: 409
                detail: No se puede suspender el inmueble porque tiene 2 visitas confirmadas en las próximas 48 horas.
                instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/status
                code: property_state_conflict
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/properties/{id}/schedules:
    parameters:
      - $ref: '#/components/parameters/PropertyId'

    get:
      tags:
        - Disponibilidad
      operationId: getPropertySchedules
      summary: Obtener las franjas horarias de una propiedad
      description: |
        Devuelve las franjas habilitadas para agendar visitas al inmueble,
        agrupadas por día de la semana.

        Cada franja es un intervalo `[start_time, end_time)` en formato
        `HH:MM` (24 h). El extremo derecho es **exclusivo**: una franja
        `09:00-12:00` y otra `12:00-15:00` son contiguas, no solapadas.
      responses:
        '200':
          description: Franjas obtenidas exitosamente
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertySchedules'
              example:
                property_id: b2c1f5a0-1a2b-4c3d-9e0f-111111111111
                timezone: America/Lima
                total_slots: 3
                days:
                  - weekday: LUNES
                    slots:
                      - id: 9a7b1c2d-3e4f-5a6b-8c7d-9e0f1a2b3c4d
                        start_time: '09:00'
                        end_time: '12:00'
                        is_active: true
                  - weekday: MARTES
                    slots: []
                  - weekday: MIERCOLES
                    slots: []
                  - weekday: JUEVES
                    slots: []
                  - weekday: VIERNES
                    slots: []
                  - weekday: SABADO
                    slots:
                      - id: 9a7b1c2d-3e4f-5a6b-8c7d-9e0f1a2b3c4f
                        start_time: '10:00'
                        end_time: '13:00'
                        is_active: true
                  - weekday: DOMINGO
                    slots: []
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Propiedad no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Disponibilidad
      operationId: replacePropertySchedules
      summary: Reemplazar las franjas horarias de una propiedad
      description: |
        Operación de **reemplazo completo** (idempotente por diseño): el cuerpo
        es la nueva configuración completa de la semana. Los bloques que no
        vienen en el cuerpo se desactivan (`is_active = false`) en lugar de
        borrarse, para conservar el historial de citas ya agendadas.

        Se eligió `PUT` y no `PATCH` porque editar un solo bloque dejaría el resto
        ambiguo. Un `PUT` con `days: []` deja el inmueble sin franjas, que es un
        estado válido: bloquea el agendamiento sin desactivar la propiedad.

        #### Reglas de validación

        | Regla | Rechazo |
        |---|---|
        | `start_time` anterior a `end_time` | `400` |
        | Duración mínima de 30 minutos | `400` |
        | Duración máxima de 8 horas | `400` |
        | Sin solapamiento entre franjas del mismo día | `400` |
        | Sin duplicar el mismo día de la semana | `400` |
        | Máximo 6 franjas por día | `400` |
        | Máximo 28 franjas por semana | `400` |

        Los horarios se interpretan en la zona horaria `America/Lima` que
        devuelve el `GET`. No se aceptan sufijos de zona ni offsets UTC en los
        valores.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertySchedulesInput'
            examples:
              semanaCompleta:
                summary: Lunes a viernes, 09:00 a 18:00
                value:
                  days:
                    - weekday: LUNES
                      slots:
                        - start_time: '09:00'
                          end_time: '12:00'
                        - start_time: '14:00'
                          end_time: '18:00'
                    - weekday: MARTES
                      slots:
                        - start_time: '09:00'
                          end_time: '18:00'
                    - weekday: MIERCOLES
                      slots:
                        - start_time: '09:00'
                          end_time: '18:00'
                    - weekday: JUEVES
                      slots:
                        - start_time: '09:00'
                          end_time: '18:00'
                    - weekday: VIERNES
                      slots:
                        - start_time: '09:00'
                          end_time: '13:00'
              sinFranjas:
                summary: Sin franjas, sin desactivar el inmueble
                value:
                  days: []
      responses:
        '200':
          description: Franjas reemplazadas exitosamente
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertySchedules'
        '400':
          description: Franjas que incumplen las reglas de rango o solapamiento
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemBadRequest'
              examples:
                rangoInvertido:
                  summary: '`start_time` posterior a `end_time`'
                  value:
                    type: https://housebroker.pe/errors/invalid-time-range
                    title: Rango horario inválido
                    status: 400
                    detail: La franja de LUNES tiene 'start_time' 18:00 posterior a 'end_time' 12:00.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/schedules
                    code: invalid_time_range
                    errors:
                      - parameter: days[0].slots[0]
                        message: "start_time' (18:00) debe ser anterior a 'end_time' (12:00)."
                duracionMinima:
                  summary: 'Franja de 15 minutos'
                  value:
                    type: https://housebroker.pe/errors/invalid-time-range
                    title: Rango horario inválido
                    status: 400
                    detail: La franja de LUNES dura 15 minutos y el mínimo es 30.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/schedules
                    code: invalid_time_range
                    errors:
                      - parameter: days[0].slots[1]
                        message: La duración de la franja (15 min) es menor que el mínimo (30 min).
                solapamiento:
                  summary: Dos franjas que se superponen en el mismo día
                  value:
                    type: https://housebroker.pe/errors/schedule-overlap
                    title: Franjas horarias solapadas
                    status: 400
                    detail: Las franjas 09:00-12:00 y 11:00-13:00 de LUNES se solapan.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/schedules
                    code: schedule_overlap
                    errors:
                      - parameter: days[0].slots[1]
                        message: Se solapa con la franja 09:00-12:00.
                formatoDeHora:
                  summary: '`start_time` fuera del patrón HH:MM'
                  value:
                    type: https://housebroker.pe/errors/invalid-time-range
                    title: Rango horario inválido
                    status: 400
                    detail: El campo 'start_time' debe usar el formato HH:MM en 24 horas.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/schedules
                    code: invalid_time_range
                    errors:
                      - parameter: days[0].slots[0].start_time
                        message: El valor '9:00 am' no cumple el formato HH:MM (patrón ^([01]\d|2[0-3]):[0-5]\d$).
                diaDuplicado:
                  summary: 'Mismo `weekday` repetido'
                  value:
                    type: https://housebroker.pe/errors/invalid-time-range
                    title: Rango horario inválido
                    status: 400
                    detail: El día 'LUNES' aparece más de una vez en la configuración semanal.
                    instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/schedules
                    code: invalid_time_range
                    errors:
                      - parameter: days[1]
                        message: El día 'LUNES' está duplicado; agrupe ambas listas de franjas en una sola entrada.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Propiedad no encontrada
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemNotFound'
        '409':
          description: |
            El inmueble tiene citas ya confirmadas en franjas que el reemplazo
            desactivaría. El cliente debe reasignar o cancelar esas visitas antes
            de aplicar el cambio.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemConflict'
              example:
                type: https://housebroker.pe/errors/schedule-has-bookings
                title: La configuración choca con visitas agendadas
                status: 409
                detail: 3 franjas a desactivar tienen visitas confirmadas.
                instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/schedules
                code: schedule_has_bookings
                errors:
                  - parameter: days[1].slots[0]
                    message: La franja 14:00-18:00 del JUEVES tiene 1 visita confirmada (cita a9c1d2e3).
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT emitido por `POST /api/v1/auth/login` y enviado como
        `Authorization: Bearer <token>`. Cubre los tres roles del MVP:
        `CLIENTE`, `AGENTE` y `ADMINISTRADOR`.

        El cliente adjunta el token en `frontend/src/services/axios.ts` leyendo
        `localStorage.hb_token`.

    refreshCookie:
        type: apiKey
        in: cookie
        name: hb_refresh_token
        description: |
          Refresh token en cookie `HttpOnly`, que es el único mecanismo con el que
          `POST /api/v1/auth/refresh` valida la sesión. El navegador la adjunta
          solo, sin que el código del cliente tenga que leerla.

          Atributos que debe fijar el backend:

          | Atributo | Valor | Motivo |
          |---|---|---|
          | `HttpOnly` | presente | Impide que `document.cookie` lo lea, que es el impacto de XSS |
          | `Secure` | presente | Solo viaja por HTTPS |
          | `SameSite` | `Strict` | Corta el envío en navegaciones originadas fuera del sitio |
          | `Path` | `/api/v1/auth` | Limita el envío a los endpoints que lo necesitan |
          | `Max-Age` | igual a `REFRESH_LIFETIME` | La sesión expira sola |

          Al no ser accesible desde JavaScript, este token **no habilita CSRF por
          sí solo** en el resto de la API: los endpoints que solo aceptan
          `Authorization: Bearer` siguen siendo inmunes porque el navegador no
          adjunta cabeceras personalizadas en un envío automático. Aun así,
          `POST /api/v1/auth/login` y `POST /api/v1/auth/refresh` sí modifican el
          estado de la sesión y deben validar un token CSRF de doble envío.
  responses:
    Unauthorized:
      description: Token ausente, vencido o con firma inválida
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemUnauthorized'
          example:
            type: https://housebroker.pe/errors/unauthorized
            title: No autenticado
            status: 401
            detail: Falta el encabezado 'Authorization' con un token Bearer válido.
            instance: /api/v1/properties/b2c1f5a0-1a2b-4c3d-9e0f-111111111111/status
            code: unauthorized
...
  parameters:
    PropertyId:
      name: id
      in: path
      required: true
      description: Identificador único del inmueble.
      schema:
        type: string
        format: uuid
...
  headers:
    XRequestId:
      description: Identificador de correlación devuelto para trazar la petición.
      schema:
        type: string
        format: uuid
  schemas:
    Property:
      type: object
      description: Representación de un inmueble del catálogo.
      required:
        - id
        - title
        - price
        - moneda
        - mode
        - address
        - property_type
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del inmueble.
          example: b2c1f5a0-1a2b-4c3d-9e0f-111111111111
        title:
          type: string
          minLength: 1
          description: Título comercial del anuncio.
          example: Departamento amoblado en Miraflores
        price:
          type: number
          format: double
          minimum: 0.0001
          description: Monto del precio en la moneda indicada por `moneda`. Siempre mayor que 0.
          example: 150000.0
        moneda:
          $ref: '#/components/schemas/Moneda'
        mode:
          $ref: '#/components/schemas/TransactionMode'
        address:
          type: string
          minLength: 1
          description: Dirección completa dentro de Lima Metropolitana.
          example: Av. Larco 456, Miraflores
        property_type:
          $ref: '#/components/schemas/PropertyType'
        is_active:
          type: boolean
          description: |
            Indica si el inmueble está visible en el catálogo. Se desactiva
            mediante eliminación lógica (`DELETE`).
          example: true
        is_favorite:
          type: boolean
          description: |
            Indica si **el usuario de la petición** guardó este inmueble en
            favoritos. Es lo que permite pintar el corazón de `PropertyCard` sin
            una llamada por inmueble.

            Depende del access token, así que el mismo catálogo devuelve
            valores distintos según quién pregunte. Todas las lecturas de este
            contrato exigen `bearerAuth` por el `security` global, así que el
            campo siempre se puede calcular; si algún día publica el
            catálogo sin token, tendría que omitirlo o mandarlo `false`.
          example: true
        created_at:
          type: string
          format: date-time
          description: Fecha y hora de registro del inmueble en formato ISO 8601.
          example: '2026-09-07T10:00:00Z'
        area_total:
          type: number
          format: double
          minimum: 0
          description: Área total del inmueble en metros cuadrados.
          example: 85.5
        area_construida:
          type: number
          format: double
          minimum: 0
          description: Área construida en metros cuadrados.
          example: 78.0
        dormitorios:
          type: integer
          minimum: 0
          description: Cantidad de dormitorios.
          example: 3
        banos:
          type: integer
          minimum: 0
          description: Cantidad de baños.
          example: 2
        estacionamientos:
          type: integer
          minimum: 0
          description: Cantidad de estacionamientos.
          example: 1
        link_galeria:
          type: string
          format: uri
          description: Enlace externo a la galería de fotografías.
          example: https://galeria.housebroker.pe/departamento-miraflores-01
        link_planos:
          type: string
          format: uri
          description: Enlace externo a los planos del inmueble.
          example: https://planos.housebroker.pe/departamento-miraflores-01.pdf
        negociable:
          type: boolean
          description: Indica si el precio admite negociación.
          example: true
        destacado:
          type: boolean
          description: Indica si el anuncio se destaca en el catálogo.
          example: false
        mantenimiento:
          type: number
          format: double
          minimum: 0
          description: Costo mensual de mantenimiento, en la moneda de `moneda`.
          example: 850.0
...
    Moneda:
      type: string
      description: Moneda en la que se exprima el precio y el mantenimiento.
      enum:
        - PEN
        - USD
      example: PEN
...
    ProblemNotFound:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required:
            - code
          properties:
            code:
              type: string
              description: Código de error estable.
              example: resource_not_found
...